수정 페이지 추가 (내힘으로...)

+++ src/main/java/com/example/demo/controller/UserArticleController.java
@@ -7,6 +7,7 @@ import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import com.example.demo.DemoApplication;
@@ -14,7 +15,11 @@ import com.example.demo.service.ArticleService;
 import com.example.demo.util.Ut;
 import com.example.demo.vo.Article;
 import com.example.demo.vo.ResultData;
+import com.example.demo.vo.Rq;
+import com.mysql.cj.Session;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -22,17 +27,25 @@ import lombok.NoArgsConstructor;
 @Controller
 public class UserArticleController {
 
+	private final DemoApplication demoApplication;
+
 	@Autowired
 	private ArticleService articleService;
+
+	UserArticleController(DemoApplication demoApplication) {
+		this.demoApplication = demoApplication;
+	}
+
+	// 로그인 체크 -> 유무 체크 -> 권한체크
 
 	// 생성(doAdd)했던 게시글'들' 출력
 
 	@RequestMapping("/user/article/getArticles")
 	@ResponseBody
-	public ResultData getArticles() {
+	public ResultData<List<Article>> getArticles() {
 
 		List<Article> articles = articleService.getArticles();
-		return ResultData.from("S-1", "Article List", articles);
+		return ResultData.from("S-1", "Article List", "게시글 리스트", articles);
 
 	}
 
@@ -40,7 +53,7 @@ public class UserArticleController {
 
 	@RequestMapping("/user/article/getArticle")
 	@ResponseBody
-	public ResultData getArticle(int id) {
+	public ResultData<Article> getArticle(int id) {
 
 		Article article = articleService.getArticleById(id);
 
@@ -48,50 +61,130 @@ public class UserArticleController {
 			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다.", id));
 		}
 
-		return ResultData.from("S-1", Ut.f("%d번 게시글입니다.", id), article);
+		return ResultData.from("S-1", Ut.f("%d번 게시글입니다.", id), "게시글 1row", article);
+	}
+
+	// 상세 페이지
+
+	@RequestMapping("/user/article/detail")
+	public String showDetail(HttpServletRequest req, Model model, int id) {
+
+		Rq rq = new Rq(req);
+
+		Article article = articleService.getForPrintArticle(rq.getLoginedMemberId(), id);
+
+		model.addAttribute("article", article);
+
+		return "user/article/detail";
+	}
+
+	// 글 리스트
+
+	@RequestMapping("/user/article/list")
+	public String showList(Model model) {
+
+		List<Article> articles = articleService.getArticles();
+
+		model.addAttribute("articles", articles);
+
+		return "user/article/list";
 	}
 
 	// 데이터 삭제
 
 	@RequestMapping("/user/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public String doDelete(HttpServletRequest req, int id) {
+
+		Rq rq = new Rq(req);
 
 		Article article = articleService.getArticleById(id);
 
-		if (article == null) {
-			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다.", id));
+		if (rq.isLogined() == false) {
+			return Ut.isReplace("F-A", "로그인 후 이용하세요", "../member/login");
 		}
 
-		articleService.deleteArticle(id);
+		if (article == null) {
+			return Ut.isHistoryBack("F-1", Ut.f("%d번 글은 없습니다.", id));
+		}
 
-		return ResultData.from("S-1", Ut.f("%d번 게시글을 지웁니다.", id), article);
+		ResultData userCanDeleteRd = articleService.userCanDelete(rq.getLoginedMemberId(), article);
+
+		if (userCanDeleteRd.isfail()) {
+			return Ut.isHistoryBack(userCanDeleteRd.getResultCode(), userCanDeleteRd.getMsg());
+		}
+
+		if (userCanDeleteRd.isSuccess()) {
+			articleService.deleteArticle(id);
+		}
+
+		return Ut.isReplace(userCanDeleteRd.getResultCode(), userCanDeleteRd.getMsg(), "../article/list");
+	}
+
+	// 수정 페이지 보여주기
+
+	@RequestMapping("/user/article/modify")
+	public String showModify(HttpSession session, Model model, int id, String title, String body) {
+
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+
+		if (session.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		}
+
+		Article article = articleService.getForPrintArticle(loginedMemberId, id);
+
+		model.addAttribute("article", article);
+
+		return "user/article/modify";
 	}
 
 	// 글 수정
 
 	@RequestMapping("/user/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
+	public ResultData doModify(HttpServletRequest req, int id, String title, String body) {
 
-		System.out.println("id" + id);
-		System.out.println("title " + title);
-		System.out.println("body " + body);
+		Rq rq = new Rq(req);
+
+		if (rq.isLogined() == false) {
+			return ResultData.from("F-A", "로그인 하세요.");
+		}
 
 		Article article = articleService.getArticleById(id);
 
+		// 게시글 유무 체크
 		if (article == null) {
-			return ResultData.from("F-1", Ut.f("%번 글은 없습니다.", id));
+			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다", id), "없는 글의 id", id);
 		}
 
-		articleService.modifyArticle(id, title, body);
+		// 수정할 수 있는 권한 체크
+		ResultData userCanModifyRd = articleService.userCanModify(rq.getLoginedMemberId(), article);
 
-		return ResultData.from("S-1", Ut.f("%d번 게시글을 수정합니다.", id), article);
+		if (userCanModifyRd.isfail()) {
+			return userCanModifyRd;
+		}
+
+		if (userCanModifyRd.isSuccess()) {
+			articleService.modifyArticle(id, title, body);
+		}
+
+		article = articleService.getArticleById(id);
+
+		return ResultData.from(userCanModifyRd.getResultCode(), userCanModifyRd.getMsg(), "수정된 글", article);
 	}
 
 	@RequestMapping("/user/article/doWrite")
 	@ResponseBody
-	public ResultData doWrite(String title, String body) {
+	public ResultData<Article> doWrite(HttpServletRequest req, String title, String body) {
+
+		Rq rq = new Rq(req);
+
+		if (rq.isLogined() == false) {
+			return ResultData.from("F-A", "로그인을 해주세요.");
+		}
 
 		if (Ut.isEmptyOrNull(title)) {
 			return ResultData.from("F-1", "제목을 입력하세요.");
@@ -101,13 +194,13 @@ public class UserArticleController {
 			return ResultData.from("F-2", "내용을 입력하세요.");
 		}
 
-		ResultData writeArticleRd = articleService.writeArticle(title, body);
+		ResultData doWriteRd = articleService.writeArticle(rq.getLoginedMemberId(), title, body);
 
-		int id = (int) writeArticleRd.getData1();
+		int id = (int) doWriteRd.getData1();
 
 		Article article = articleService.getArticleById(id);
 
-		return ResultData.from(writeArticleRd.getResultCode(), writeArticleRd.getMsg(), article);
+		return ResultData.newData(doWriteRd, "새로 작성된 게시글", article);
 
 	}
 }
